requirements.txt
django
django-taggit
markdown
psycopg2
Pillow
wagtail
wagtailaltgenerator
djangorestframework
wagtail-headless-preview
django-cors-headers